deprecate Insomnia AI and move spec actions to file route
