+++ packages/insomnia/src/ui/routes/actions.tsx
@@ -1,25 +1,11 @@
-import path from 'node:path';
-
-import { type IRuleResult } from '@stoplight/spectral-core';
-import { generate, runTests, type Test } from 'insomnia-testing';
-import type { TestResults } from 'insomnia-testing/src/run/entities';
 import { type ActionFunction, redirect } from 'react-router';
 
-import { parseApiSpec, resolveComponentSchemaRefs } from '../../common/api-specs';
-import { ACTIVITY_DEBUG, getAIServiceURL } from '../../common/constants';
 import { database } from '../../common/database';
-import { importResourcesToWorkspace, scanResources, type ScanResult } from '../../common/import';
-import { generateId } from '../../common/misc';
 import * as models from '../../models';
 import { EnvironmentType } from '../../models/environment';
 import { getById, update } from '../../models/helpers/request-operations';
-import { isGitProject } from '../../models/project';
-import { isRequest, type Request } from '../../models/request';
 import { isRequestGroup, isRequestGroupId } from '../../models/request-group';
 import { isRequestGroupMeta } from '../../models/request-group-meta';
-import type { UnitTest } from '../../models/unit-test';
-import { getSendRequestCallback } from '../../network/unit-test-feature';
-import { insomniaFetch } from '../../ui/insomniaFetch';
 import { invariant } from '../../utils/invariant';
 import { SegmentEvent } from '../analytics';
 
@@ -37,343 +23,6 @@ export function safeToUseInsomniaFileName(fileName: string) {
 export function safeToUseInsomniaFileNameWithExt(fileName: string) {
   return `${safeToUseInsomniaFileName(fileName)}.yaml`;
 }
-
-// Api Spec
-export const updateApiSpecAction: ActionFunction = async ({ request, params }) => {
-  const { workspaceId } = params;
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-  const formData = await request.formData();
-  const contents = formData.get('contents');
-  const fromSync = Boolean(formData.get('fromSync'));
-
-  invariant(typeof contents === 'string', 'Contents is required');
-
-  const apiSpec = await models.apiSpec.getByParentId(workspaceId);
-
-  invariant(apiSpec, 'API Spec not found');
-  await database.update(
-    {
-      ...apiSpec,
-      modified: Date.now(),
-      created: fromSync ? Date.now() : apiSpec.created,
-      contents,
-    },
-    fromSync,
-  );
-
-  return null;
-};
-
-export const generateCollectionFromApiSpecAction: ActionFunction = async ({ params }) => {
-  const { organizationId, projectId, workspaceId } = params;
-
-  invariant(typeof projectId === 'string', 'Project ID is required');
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-
-  const project = await models.project.getById(projectId);
-  invariant(project, 'Project not found');
-
-  const apiSpec = await models.apiSpec.getByParentId(workspaceId);
-  invariant(apiSpec, 'No API Specification was found');
-
-  const workspace = await models.workspace.getById(workspaceId);
-
-  invariant(workspace, 'Workspace not found');
-
-  const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspaceId);
-
-  const isLintError = (result: IRuleResult) => result.severity === 0;
-
-  const gitRepositoryId = isGitProject(project) ? project.gitRepositoryId : workspaceMeta?.gitRepositoryId;
-
-  const rulesetPath = gitRepositoryId
-    ? path.join(window.app.getPath('userData'), `version-control/git/${gitRepositoryId}/other/.spectral.yaml`)
-    : '';
-
-  const { diagnostics, error } = await window.main.lintSpec({ documentContent: apiSpec.contents, rulesetPath });
-  if (error) {
-    throw error;
-  }
-  const results = diagnostics?.filter(isLintError);
-  if (apiSpec.contents && results && results.length) {
-    throw new Error('Error Generating Configuration');
-  }
-
-  await scanResources([
-    {
-      contentStr: apiSpec.contents,
-    },
-  ]);
-
-  await importResourcesToWorkspace({
-    workspaceId,
-  });
-
-  return redirect(`/organization/${organizationId}/project/${projectId}/workspace/${workspaceId}/${ACTIVITY_DEBUG}`);
-};
-
-export const generateCollectionAndTestsAction: ActionFunction = async ({ params }) => {
-  const { organizationId, projectId, workspaceId } = params;
-
-  invariant(typeof organizationId === 'string', 'Organization ID is required');
-  invariant(typeof projectId === 'string', 'Project ID is required');
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-
-  const apiSpec = await models.apiSpec.getByParentId(workspaceId);
-
-  invariant(apiSpec, 'API Spec not found');
-
-  const workspace = await models.workspace.getById(workspaceId);
-
-  invariant(workspace, 'Workspace not found');
-
-  const workspaceMeta = await models.workspaceMeta.getOrCreateByParentId(workspaceId);
-
-  const gitRepositoryId = workspaceMeta?.gitRepositoryId;
-  const isLintError = (result: IRuleResult) => result.severity === 0;
-  const rulesetPath = gitRepositoryId
-    ? path.join(window.app.getPath('userData'), `version-control/git/${gitRepositoryId}/other/.spectral.yaml`)
-    : '';
-
-  const { diagnostics, error } = await window.main.lintSpec({ documentContent: apiSpec.contents, rulesetPath });
-  if (error) {
-    throw error;
-  }
-  const results = diagnostics?.filter(isLintError);
-  if (apiSpec.contents && results && results.length) {
-    throw new Error('Error Generating Configuration');
-  }
-
-  const resources = await scanResources([
-    {
-      contentStr: apiSpec.contents,
-    },
-  ]);
-
-  const allRequestsFromResources = resources.reduce(
-    (accumulator, scanResult) => accumulator.concat(scanResult.requests ?? []),
-    [] as NonNullable<ScanResult['requests']>,
-  );
-
-  const aiGeneratedRequestGroup = await models.requestGroup.create({
-    name: 'AI Generated Requests',
-    parentId: workspaceId,
-  });
-
-  const requests =
-    allRequestsFromResources.filter(isRequest).map(request => {
-      return {
-        ...request,
-        _id: generateId(models.request.prefix),
-        parentId: aiGeneratedRequestGroup._id,
-      };
-    }) || [];
-
-  await Promise.all(requests.map(request => models.request.create(request)));
-
-  const aiTestSuite = await models.unitTestSuite.create({
-    name: 'AI Generated Tests',
-    parentId: workspaceId,
-  });
-
-  const spec = parseApiSpec(apiSpec.contents);
-
-  const getMethodInfo = (request: Request) => {
-    try {
-      const specPaths = Object.keys(spec.contents?.paths) || [];
-
-      const pathMatches = specPaths.filter(path => request.url.endsWith(path));
-
-      const closestPath = pathMatches.sort((a, b) => {
-        return a.length - b.length;
-      })[0];
-
-      const methodInfo = spec.contents?.paths[closestPath][request.method.toLowerCase()];
-
-      return methodInfo;
-    } catch (error) {
-      console.log(error);
-      return undefined;
-    }
-  };
-
-  const tests: Partial<UnitTest>[] = requests.map(request => {
-    return {
-      name: `Test: ${request.name}`,
-      code: '',
-      parentId: aiTestSuite._id,
-      requestId: request._id,
-    };
-  });
-
-  const total = tests.length;
-  let progress = 0;
-
-  // @TODO Investigate the defer API for streaming results.
-  const progressStream = new TransformStream();
-  const writer = progressStream.writable.getWriter();
-
-  writer.write({
-    progress,
-    total,
-  });
-
-  for (const test of tests) {
-    async function generateTest() {
-      try {
-        const request = requests.find(r => r._id === test.requestId);
-        if (!request) {
-          throw new Error('Request not found');
-        }
-
-        const user = await models.userSession.getOrCreate();
-        const sessionId = user.id;
-
-        const methodInfo = resolveComponentSchemaRefs(spec, getMethodInfo(request));
-        const response = await insomniaFetch<{ test: { requestId: string } }>({
-          method: 'POST',
-          origin: getAIServiceURL(),
-          path: '/v1/generate-test',
-          sessionId,
-          data: {
-            teamId: organizationId,
-            request: requests.find(r => r._id === test.requestId),
-            methodInfo,
-          },
-        });
-
-        const aiTest = response.test;
-
-        await models.unitTest.create({ ...aiTest, parentId: aiTestSuite._id, requestId: test.requestId });
-        writer.write({
-          progress: ++progress,
-          total,
-        });
-      } catch (err) {
-        console.log(err);
-        writer.write({
-          progress: ++progress,
-          total,
-        });
-      }
-    }
-    generateTest();
-  }
-
-  return progressStream;
-};
-
-export const generateTestsAction: ActionFunction = async ({ params }) => {
-  const { organizationId, projectId, workspaceId } = params;
-
-  invariant(typeof organizationId === 'string', 'Organization ID is required');
-  invariant(typeof projectId === 'string', 'Project ID is required');
-  invariant(typeof workspaceId === 'string', 'Workspace ID is required');
-
-  const apiSpec = await models.apiSpec.getByParentId(workspaceId);
-
-  invariant(apiSpec, 'API Spec not found');
-
-  const workspace = await models.workspace.getById(workspaceId);
-
-  invariant(workspace, 'Workspace not found');
-
-  const workspaceDescendants = await database.withDescendants(workspace);
-
-  const requests = workspaceDescendants.filter(isRequest);
-
-  const aiTestSuite = await models.unitTestSuite.create({
-    name: 'AI Generated Tests',
-    parentId: workspaceId,
-  });
-
-  const tests: Partial<UnitTest>[] = requests.map(request => {
-    return {
-      name: `Test: ${request.name}`,
-      code: '',
-      parentId: aiTestSuite._id,
-      requestId: request._id,
-    };
-  });
-
-  const total = tests.length;
-  let progress = 0;
-  // @TODO Investigate the defer API for streaming results.
-  const progressStream = new TransformStream();
-  const writer = progressStream.writable.getWriter();
-
-  writer.write({
-    progress,
-    total,
-  });
-
-  async function generateTests() {
-    async function generateTest(test: Partial<UnitTest>) {
-      const user = await models.userSession.getOrCreate();
-      const sessionId = user.id;
-      try {
-        const response = await insomniaFetch<{ test: { requestId: string } }>({
-          method: 'POST',
-          origin: getAIServiceURL(),
-          path: '/v1/generate-test',
-          sessionId,
-          data: {
-            teamId: organizationId,
-            request: requests.find(r => r._id === test.requestId),
-          },
-        });
-
-        const aiTest = response.test;
-
-        await models.unitTest.create({ ...aiTest, parentId: aiTestSuite._id, requestId: test.requestId });
-
-        writer.write({
-          progress: ++progress,
-          total,
-        });
-      } catch (err) {
-        console.log(err);
-        writer.write({
-          progress: ++progress,
-          total,
-        });
-      }
-    }
-
-    for (const test of tests) {
-      await generateTest(test);
-    }
-  }
-
-  generateTests();
-
-  return progressStream;
-};
-
-export const accessAIApiAction: ActionFunction = async ({ params }) => {
-  const { organizationId } = params;
-
-  invariant(typeof organizationId === 'string', 'Organization ID is required');
-
-  try {
-    const user = await models.userSession.getOrCreate();
-    const sessionId = user.id;
-    const response = await insomniaFetch<{ enabled: boolean }>({
-      method: 'POST',
-      origin: getAIServiceURL(),
-      path: '/v1/access',
-      sessionId,
-      data: {
-        teamId: organizationId,
-      },
-    });
-    return {
-      enabled: response.enabled,
-    };
-  } catch (err) {
-    return { enabled: false };
-  }
-};
 
 export const createEnvironmentAction: ActionFunction = async ({ params, request }) => {
   const { workspaceId } = params;
@@ -393,6 +42,7 @@ export const createEnvironmentAction: ActionFunction = async ({ params, request 
 
   return environment;
 };
+
 export const updateEnvironment: ActionFunction = async ({ request, params }) => {
   const { workspaceId } = params;
   invariant(typeof workspaceId === 'string', 'Workspace ID is required');
